fix(control): change product id data type of commands.

+++ AquaEngine.API/Control/Domain/Model/Commands/DecreaseQuantityCommand.cs
@@ -1,3 +1,3 @@
 ﻿namespace AquaEngine.API.Control.Domain.Model.Commands;
 
-public record DecreaseQuantityCommand(int Quantity, long ProductId);
+public record DecreaseQuantityCommand(int Quantity, int ProductId);
